Refactor BVH construction to support a maximum depth
This commit refactors the code for building the BVH (Bounding Volume Hierarchy) to support a maximum depth. Previously, the BVH construction would continue indefinitely, potentially leading to performance issues. With this change, a maximum depth parameter is introduced, allowing developers to control the depth of the BVH tree. This improves performance and memory usage by limiting the recursion depth.

The `buildBVHNode` function now takes an additional `maxDepth` parameter, which determines the maximum depth allowed for the BVH tree. If the current depth exceeds the maximum depth or if the number of triangles is less than or equal to 2, the function stops the recursion and creates a leaf node. Otherwise, it continues to split the triangles and build the BVH tree.

This change also updates the `ConvertObjectsToBVH` and `object.BuildBVH` functions to pass the `maxDepth` parameter to the `buildBVHNode` function.

Note: The `maxDepth` constant is introduced in this commit and set to 12.